requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==35.0.0
Django==3.2.7
django-cors-headers==3.10.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.2
